All: rename tilemap to tileset
The actual concept I was trying to model with the tilemaps is a tileset,
a set of tiles that can be used for multiple things like characters or
the actual map.

This change was necessary as the next commits will actually include a
tilemap object that will be in charge of rendering the map itself using
the scene.png tileset.