Game: more dirt variation and more terrain visible
Before, by just following the mario level layout, there was too much air
at the top of the level. By adding another dirt layer at the bottom and
removing a layer of air at the top the player character is more centered
in the screen, which in my opinion looks much better.

After adding a layer of dirt, more dirt variations have been added in
order to break the monotony of the texture.