Game: first update, then render
The player character jittered on the screen. In theory, as the player
position and the camera position are the same, the player character
should not move at all in screen position.

The issue was that the function `gfx_frame_begin` function sets the
global camera position for the frame, after that the player position was
updated and the camera position variable aswell, but not the camera
position used for rendering.

This was fixed by first updating the game state and later starting the
rendering process.